Organise class into suitable java packages
Organise java class into the following packages:
- bot.ui
- bot.command
- bot.task
- bot.service
- bot.util
- bot.exception

This allows the project structure to be more organised. Which is crucial as project size grows.
