+++ src/main/java/Main.java
@@ -1,6 +1,8 @@
+import bot.Bot;
+
 public class Main {
     public static void main(String[] args) {
-        // Init a Bot to begin program
+        // Init a bot.Bot to begin program
         Bot bot = new Bot("Lovely", "data/taskData.txt");
         bot.run();
     }
