+++ src/main/java/bot/command/MarkTaskCommand.java
@@ -1,3 +1,11 @@
+package bot.command;
+
+import bot.service.FileServices;
+import bot.exception.InvalidCommandException;
+import bot.task.TaskList;
+import bot.ui.Ui;
+import bot.task.Task;
+
 public class MarkTaskCommand extends Command {
     private final String[] commandInfo;
 
@@ -12,7 +20,7 @@ public class MarkTaskCommand extends Command {
             if (commandInfo.length != 2 || !commandInfo[1].matches("\\d+")) {
                 throw new InvalidCommandException(
                         "Please ensure command is in this format: " +
-                                "mark <Task Index>");
+                                "mark <bot.task.Task Index>");
             }
 
             int index = Integer.parseInt(commandInfo[1]);
