+++ src/main/java/bot/task/Todo.java
@@ -1,3 +1,5 @@
+package bot.task;
+
 public class Todo extends Task {
     public Todo(String taskName) {
         super(taskName);
