+++ src/main/java/bot/util/DateTimeUtils.java
@@ -1,3 +1,5 @@
+package bot.util;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
